split files to make it norminette-proof and fixed other norminette issues
