+++ Makefile
@@ -16,7 +16,9 @@ SOURCE_FILES	= main.c check_for_builtins.c exact_match.c exit_handler.c \
 				  parser.c cmd_list_utils.c cmd_make_utils.c cmd_free_utils.c \
 				  cmd_execute.c cmd_path_utils.c signal_handler.c \
 				  env_list_utils.c env_make_utils.c env_export_utils.c \
-				  remove_quotes.c parameter_expansion.c
+					remove_quotes.c parameter_expansion.c builtin_cd.c \
+					builtin_export.c builtin_functions.c lexer_handle_quotes.c \
+					parameter_expansion_utils.c lexer_handle_make_var.c	
 
 SOURCES_DIR		= src
 SOURCES			= $(addprefix $(SOURCES_DIR)/, $(SOURCE_FILES))
