+++ src/handle_meta_char.c
@@ -6,7 +6,7 @@
 /*   By: pclaus <pclaus@student.s19.be>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/18 21:31:35 by pclaus            #+#    #+#             */
-/*   Updated: 2024/05/28 14:13:17 by pclaus           ###   ########.fr       */
+/*   Updated: 2024/07/11 19:36:06 by pclaus           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ void	handle_meta_char(t_lexeme *lexeme, char *line, int *index)
 		(*index)++;
 		iter++;
 		if (!is_meta_character(line[*index]) || (line[*index] != line[*index
-				- 1]) || iter == 2 || line[*index] == '\0')
+					- 1]) || iter == 2 || line[*index] == '\0')
 		{
 			reset_lexer_state(lexeme, START);
 			break ;
