+++ src/lexer_handler_functions.c
@@ -6,67 +6,11 @@
 /*   By: pclaus <pclaus@student.s19.be>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/20 10:20:46 by pclaus            #+#    #+#             */
-/*   Updated: 2024/07/07 14:47:02 by pclaus           ###   ########.fr       */
+/*   Updated: 2024/07/11 19:46:45 by pclaus           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
-
-void	handle_single_quotes(t_lexeme *lexeme, char *line, int *index)
-{
-//	(*index)++;
-	while (line[*index] != '\0' && lexeme->lexing_state == SQ)
-	{
-		ft_strjoin_char(&lexeme->string, (line[*index]));
-		(*index)++;
-		if (line[*index] == '\'')
-		{
-			ft_strjoin_char(&lexeme->string, (line[*index]));
-			reset_lexer_state(lexeme, START);
-			(*index)++;
-		}
-	}
-}
-
-void	handle_var_single(t_lexeme *lexeme, char *line, int **index)
-{
-	lexeme->lexing_state = VAR_SINGLE;
-	ft_strjoin_char(&lexeme->string, (line[**index]));
-	while (line[**index] != '\0' && lexeme->lexing_state == VAR_SINGLE)
-	{
-		(**index)++;
-		ft_strjoin_char(&lexeme->string, (line[**index]));
-		if (line[**index] == '\'')
-		{
-			reset_lexer_state(lexeme, START);
-			(**index)++;
-		}
-	}
-}
-
-void	handle_var_double(t_lexeme *lexeme, char *line, int **index)
-{
-	lexeme->lexing_state = VAR_DOUBLE;
-	ft_strjoin_char(&lexeme->string, (line[**index]));
-	while (line[**index] != '\0' && lexeme->lexing_state == VAR_DOUBLE)
-	{
-		(**index)++;
-		ft_strjoin_char(&lexeme->string, (line[**index]));
-		if (line[**index] == '"')
-		{
-			reset_lexer_state(lexeme, START);
-			(**index)++;
-		}
-	}
-}
-
-void	handle_var_make(t_lexeme *lexeme, char *line, int **index)
-{
-	if (line[**index] == '\'' && line[**index - 1] == '=')
-		handle_var_single(lexeme, line, index);
-	else if (line[**index] == '"' && line[**index - 1] == '=')
-		handle_var_double(lexeme, line, index);
-}
 
 void	handle_unquoted(t_lexeme *lexeme, char *line, int *index)
 {
@@ -74,7 +18,8 @@ void	handle_unquoted(t_lexeme *lexeme, char *line, int *index)
 	{
 		ft_strjoin_char(&lexeme->string, (line[*index]));
 		(*index)++;
-		if ((line[*index] == '\'' || line[*index] == '"') && line[*index - 1] == '=')
+		if ((line[*index] == '\'' || line[*index] == '"') && line[*index
+				- 1] == '=')
 		{
 			handle_var_make(lexeme, line, &index);
 			break ;
@@ -86,11 +31,7 @@ void	handle_unquoted(t_lexeme *lexeme, char *line, int *index)
 		}
 		if (!is_regular_character(line[*index]) && (line[*index] != '_'
 				|| line[*index] != '\0') && (line[*index] != '$'))
-		{
 			reset_lexer_state(lexeme, START);
-//			if (line[*index] != '\0')
-//				(*index)++;
-		}
 	}
 }
 
@@ -101,22 +42,6 @@ void	handle_space(t_lexeme *lexeme, char *line, int *index)
 		(*index)++;
 		if (line[*index] != ' ')
 			lexeme->lexing_state = START;
-	}
-}
-
-void	handle_double_quotes(t_lexeme *lexeme, char *line, int *index)
-{
-//	(*index)++;
-	while (line[*index] != '\0' && lexeme->lexing_state == DQ)
-	{
-		ft_strjoin_char(&lexeme->string, (line[*index]));
-		(*index)++;
-		if (line[*index] == '"')
-		{
-			ft_strjoin_char(&lexeme->string, (line[*index]));
-			reset_lexer_state(lexeme, START);
-			(*index)++;
-		}
 	}
 }
