+++ src/tokenizer.c
@@ -6,21 +6,12 @@
 /*   By: pclaus <pclaus@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/14 15:37:56 by pclaus            #+#    #+#             */
-/*   Updated: 2024/07/04 16:27:11 by pclaus           ###   ########.fr       */
+/*   Updated: 2024/07/12 08:55:09 by pclaus           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
-/*
-void	tokenizer(char *line)
-{
-	t_token	*head;
 
-	head = NULL;
-	add_token_to_end(&head, create_token(line));
-	print_list(&head);	
-}
-*/
 t_token	*create_token(char *string)
 {
 	t_token	*new_token;
@@ -48,26 +39,28 @@ void	add_token_to_end(t_token **head, t_token *new_token)
 		iter = iter->next;
 	iter->next = new_token;
 }
-
+/*
 void	print_list(t_token **token)
 {
-	int iter;
+	int	iter;
 
 	iter = 0;
 	while (token != NULL && (*token)->str != NULL)
 	{
-		printf("#%d: The value in the node from the tokenizer is: %s ~~ The tag is: %d\n", iter, (*token)->str, (*token)->tag);
+		printf("#%d: The value in the node from the tokenizer is:\
+				%s ~~ The tag is: %d\n", iter, (*token)->str, (*token)->tag);
 		iter++;
 		if ((*token)->next != NULL)
 			*token = (*token)->next;
 		else
-		 break ;
+			break ;
 	}
 }
 
+*/
 void	free_tokens(t_token **tokens)
 {
-	t_token *free_me;
+	t_token	*free_me;
 
 	if (!tokens || !(*tokens))
 		return ;
@@ -81,3 +74,4 @@ void	free_tokens(t_token **tokens)
 	}
 	*tokens = NULL;
 }
+
